+++ source/extensions/lightspeed.trex.capture.core.shared/lightspeed/trex/capture/core/shared/tests/unit/__init__.py
@@ -1,5 +1,5 @@
 """
-* SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+* SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 * SPDX-License-Identifier: Apache-2.0
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,8 +14,3 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 """
-
-from .e2e.test_setup import TestOnLoadEvent
-from .unit.test_setup import TestSetup
-
-__all__ = ["TestOnLoadEvent", "TestSetup"]